#5383: [Falcon7b] Add perplexity tests to new pipeline and restructure pytests to invoke with filenames
