+++ tests/scripts/t3000/run_t3000_perplexity_tests.sh
@@ -7,9 +7,9 @@ run_t3000_perplexity_tests() {
   start_time=$(date +%s)
 
   echo "LOG_METAL: Running run_t3000_perplexity_tests"
-  # Insert Tests here
-  
-  echo "Hello World"
+
+  # Falcon7B perplexity tests
+  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto models/demos/falcon7b/tests/perplexity/test_perplexity_falcon.py --timeout=1500 ; fail+=$?
 
   # Record the end time
   end_time=$(date +%s)
